fix(typing): corrige type hints e verificacoes de tipo em link_layer, host e epr
4.1 - purification_calculator em link_layer.py: parametros f1 e f2 corrigidos de int para float, pois representam fidelidades (valores entre 0 e 1). Docstring atualizada.
4.2 - Host.add_connection em host.py: substituido type(x) != int por not isinstance(x, int), aceitando subclasses de int como numpy.int64 usadas pelo NetworkX.
4.3 - Host.info() em host.py: comparacao self.routing_table == None substituida por self.routing_table is None, seguindo a convencao PEP 8.
4.4 - Epr.__init__ em epr.py: epr_id anotado como Union[int, tuple] para refletir os dois usos reais: int em physical_layer e tuple (alice_id, bob_id) em link_layer e network_layer.

+++ quantumnet/layers/link_layer.py
@@ -105,13 +105,13 @@ class LinkLayer:
             alice, bob, on_complete=on_heralding_done
         )
 
-    def purification_calculator(self, f1: int, f2: int, purification_type: int) -> float:
+    def purification_calculator(self, f1: float, f2: float, purification_type: int) -> float:
         """
         Purification formula calculation.
 
         Args:
-            f1 (int): Fidelity of the first EPR.
-            f2 (int): Fidelity of the second EPR.
+            f1 (float): Fidelity of the first EPR.
+            f2 (float): Fidelity of the second EPR.
             purification_type (int): Chosen formula (1 - Default, 2 - BBPSSW Protocol, 3 - DEJMPS Protocol).
 
         Returns:
